+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.6.2] UNRELEASED
+### Added
+- `dataclassdict` decorator.
+
 ### Fixed
 - Typing for `function_alias`, `deprecated_alias` and `deprecated_function` decorators.